+++ src/allotropy/parsers/agilent_tapestation_analysis/constants.py
@@ -1,24 +1,13 @@
-from allotropy.allotrope.models.shared.definitions.custom import (
-    TQuantityValueKiloDalton,
-    TQuantityValueNumber,
-    TQuantityValueSecondTime,
-)
-
 BRAND_NAME = "TapeStation"
 DETECTION_TYPE = "fluorescence"
 DEVICE_TYPE = "electrophoresis device"
 PRODUCT_MANUFACTURER = "Agilent"
 SOFTWARE_NAME = "TapeStation Analysis Software"
 
-NO_SCREEN_TAPE_ID_MATCH = "The ScreenTape ID associated with the sample {} does not match any ScreenTape element."
-
-UNIT_CLASSES = type[
-    TQuantityValueKiloDalton | TQuantityValueNumber | TQuantityValueSecondTime
-]
-UNIT_CLASS_LOOKUP: dict[str, UNIT_CLASSES] = {
-    "nt": TQuantityValueNumber,
-    "bp": TQuantityValueNumber,
-    "kD": TQuantityValueKiloDalton,
+UNIT_CLASS_LOOKUP: dict[str, str] = {
+    "nt": "#",
+    "bp": "#",
+    "kD": "kDa",
 }
 NON_CALCULATED_DATA_TAGS_SAMPLE = [
     "Comment",