+++ src/allotropy/parsers/beckman_vi_cell_xr/vi_cell_xr_structure.py
@@ -1,26 +1,22 @@
 from __future__ import annotations
 
-import re
-
 from allotropy.allotrope.schema_mappers.adm.cell_counting.benchling._2023._11.cell_counting import (
     Data,
     Measurement,
     MeasurementGroup,
     Metadata,
 )
-from allotropy.exceptions import AllotropeConversionError
+from allotropy.exceptions import (
+    AllotropeConversionError,
+)
 from allotropy.parsers.beckman_vi_cell_xr.constants import (
     DEFAULT_ANALYST,
-    DEFAULT_VERSION,
     MODEL_NUMBER,
-    MODEL_RE,
     SOFTWARE_NAME,
-    XrVersion,
 )
 from allotropy.parsers.beckman_vi_cell_xr.vi_cell_xr_reader import ViCellData
-from allotropy.parsers.utils.pandas import map_rows, SeriesData
+from allotropy.parsers.utils.pandas import SeriesData
 from allotropy.parsers.utils.uuids import random_uuid_str
-from allotropy.parsers.utils.values import assert_not_none
 
 
 def _create_measurement_group(data: SeriesData) -> MeasurementGroup:
@@ -55,35 +51,18 @@ def _create_measurement_group(data: SeriesData) -> MeasurementGroup:
 
 
 def create_data(reader_data: ViCellData, file_name: str) -> Data:
-    serial_number_str = reader_data.file_info[str, "serial"]
-    try:
-        serial_number = serial_number_str[serial_number_str.rindex(":") + 1 :].strip()
-    except ValueError:
-        serial_number = None
-
-    match = re.match(
-        MODEL_RE,
-        reader_data.file_info[str, "model"],
-        flags=re.IGNORECASE,
-    )
-    try:
-        version_str = assert_not_none(match).groupdict()["version"]
-        version_str = ".".join(version_str.split(".")[0:2])
-        version = XrVersion(version_str)
-    except (AttributeError, AllotropeConversionError):
-        version = DEFAULT_VERSION
-    except ValueError as e:
-        msg = f"Invalid Beckman VI-Cell XR version: {version_str}"
-        raise AllotropeConversionError(msg) from e
+    if not reader_data.data:
+        msg = "Cannot parse ASM from empty file."
+        raise AllotropeConversionError(msg)
 
     metadata = Metadata(
         device_type="brightfield imager (cell counter)",
         detection_type="brightfield",
         model_number=MODEL_NUMBER,
-        equipment_serial_number=serial_number,
+        equipment_serial_number=reader_data.serial_number,
         software_name=SOFTWARE_NAME,
-        software_version=version.value,
+        software_version=reader_data.version.value,
         file_name=file_name,
     )
 
-    return Data(metadata, map_rows(reader_data.data, _create_measurement_group))
+    return Data(metadata, [_create_measurement_group(row) for row in reader_data.data])