+++ src/allotropy/parsers/thermo_fisher_nanodrop_eight/nanodrop_eight_structure.py
@@ -11,6 +11,7 @@ from allotropy.allotrope.schema_mappers.adm.spectrophotometry.benchling._2023._1
     DataSource,
     Measurement,
     MeasurementGroup,
+    MeasurementType,
     Metadata,
     ProcessedData,
     ProcessedDataFeature,
@@ -81,6 +82,7 @@ class SpectroscopyRow:
                 continue
             measurements.append(
                 Measurement(
+                    type_=MeasurementType.ULTRAVIOLET_ABSORBANCE,
                     identifier=random_uuid_str(),
                     absorbance=absorbance,
                     detector_wavelength_setting=wavelength,
@@ -148,15 +150,14 @@ def create_data(data: pd.DataFrame, file_name: str) -> Data:
         ),
         measurement_groups=[
             MeasurementGroup(
-                _measurement_time=row.timestamp,
+                measurement_time=row.timestamp,
                 analyst=row.analyst,
                 experiment_type=row.experiment_type,
                 measurements=row.measurements,
             )
             for row in rows
         ],
-        # NOTE: in current implementation, calculated data is reported at global level for some
-        # reason
+        # NOTE: in current implementation, calculated data is reported at global level for some reason.
         # TODO(nstender): should we move this inside of measurements?
         calculated_data=[item for row in rows for item in row.calculated_data],
     )