+++ tests/parsers/novabio_flex2/novabio_flex2_structure_test.py
@@ -72,7 +72,7 @@ def test_create_analyte() -> None:
 
 @pytest.mark.short
 def test_create_invalid_analyte() -> None:
-    expected_regex_raw = "Unrecognized analyte name: 'FAKE'. Only ['Ca++', 'Gln', 'Glu', 'Gluc', 'HCO3', 'K+', 'Lac', 'NH4+', 'Na+'] are supported."
+    expected_regex_raw = "Unrecognized analyte name: 'FAKE'. Expecting one of ['Ca++', 'Gln', 'Glu', 'Gluc', 'HCO3', 'K+', 'Lac', 'NH4+', 'Na+']."
     expected_regex = re.escape(expected_regex_raw)
     with pytest.raises(AllotropeConversionError, match=expected_regex):
         Analyte.create("FAKE", 100)
@@ -96,7 +96,7 @@ def test_create_sample() -> None:
 
     assert sample.identifier == "BP_R10_KP_008_D0"
     assert sample.sample_type == "Spent Media"
-    assert sample.measurement_time == "2022-06-24T14:34:52"
+    assert sample.measurement_time == "2022-06-24 14:34:52"
     assert sample.batch_identifier == "KP_008"
     assert sorted(sample.analytes) == sorted(
         [