+++ CHANGELOG.md
@@ -8,13 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Add plate id field to biorad bioplex
-- Add luminescence point detection, absorption spectrum detection, fluorescence emission detection measurement extension to Spectrophotometry BENCHLING/2023/12 schema
-- Add BMG MARS adapter
 
 ### Fixed
-- Handle comma as decimal place in float conversion utilities.
-- Add error for multiple read modes in Agilent Gen5
 
 ### Changed
 
@@ -24,11 +19,40 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.1.47] - 2024-08-13
+
+### Fixed
+
+- Fix mis-reporting some analyte units in Roche Cedex Bioht parser
+
+## [0.1.46] - 2024-08-13
+
+### Added
+
+- Add plate id field to biorad bioplex
+- Add luminescence point detection, absorption spectrum detection, fluorescence emission detection measurement extension to Spectrophotometry BENCHLING/2023/12 schema
+- Add BMG MARS adapter
+- Added plate-reader REC/2024/06 schema
+
+### Fixed
+
+- Handle comma as decimal place in float conversion utilities
+- Raise AllotropeConversionError on missing Results section in Agilent Gen5
+- Add error for multiple read modes in Agilent Gen5
+- Cast data to str before using in AppBio QuantStudio parser
+- Simplify VI-Cell XR text file reader, removing bug in pivot operation
+- Fix edge case where there are multiple measurements for a property in Roche Cedex Bioht
+
+### Changed
+
+- Only return AllotropeConversionError when there is a problem with input data that we expect, add other errors for unexpected problems.
+
 ## [0.1.45] - 2024-08-01
 
 ### Changed
 
 - Pandas version updated to 2.2.0 to have calamine engine
+- Updated Roche Cedex Bioht adapter to work with the Solution Analyzer ASM schema
 
 ## [0.1.44] - 2024-07-30